requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
jinja2>=3.1.0,<4.0.0
pydantic>=2.8.0,<3.0.0
pydantic-settings>=2.4.0,<3.0.0
PyYAML>=6.0.0,<7.0.0
python-multipart>=0.0.9,<1.0.0
pdfplumber>=0.11.0,<1.0.0
python-docx>=1.1.0,<2.0.0
pytesseract>=0.3.10,<1.0.0
pdf2image>=1.17.0,<2.0.0
playwright>=1.48.0,<2.0.0
readability-lxml>=0.8.1,<1.0.0
beautifulsoup4>=4.12.0,<5.0.0
openai>=1.50.0,<2.0.0
